stage_02

+++ 第一阶段模块二/code/MyPhone/SamePackage.java
@@ -7,7 +7,7 @@ public abstract class SamePackage {
         return cost;
     }
 
-    public void setCost(float cost) {
+    public void setCost(double cost) {
         this.cost = cost;
     }
